Move main.go inside command/command.go
