+++ cmd/dogstatsd/main_nix.go
@@ -12,6 +12,7 @@ import (
 	_ "net/http/pprof"
 	"os"
 
+	"github.com/DataDog/datadog-agent/cmd/dogstatsd/command"
 	"github.com/DataDog/datadog-agent/pkg/util/flavor"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
@@ -21,7 +22,7 @@ const defaultLogFile = "/var/log/datadog/dogstatsd.log"
 func main() {
 	flavor.SetFlavor(flavor.Dogstatsd)
 
-	if err := MakeRootCommand(defaultLogFile).Execute(); err != nil {
+	if err := command.MakeRootCommand(defaultLogFile).Execute(); err != nil {
 		log.Error(err)
 		os.Exit(-1)
 	}
